format_iso_8601 {parsedate}	R Documentation

Format date and time according to ISO 8601

Description

Format a date in a fixed format that is ISO 8601 valid, and can be used to compare dates as character strings. It converts the date(s) to UTC.

Usage

format_iso_8601(date)

Arguments

date
The date(s) to format.

Value

Character vector of formatted dates.

Examples

format_iso_8601(parse_iso_8601("2013-02-08"))
format_iso_8601(parse_iso_8601("2013-02-08 09:34:00"))
format_iso_8601(parse_iso_8601("2013-02-08 09:34:00+01:00"))
format_iso_8601(parse_iso_8601("2013-W06-5"))
format_iso_8601(parse_iso_8601("2013-039"))

[Package parsedate version 1.1.3 Index]