parse_date {parsedate}	R Documentation

Parse date from any format

Description

Recognize and parse dates from a wide range of formats. The current algorithm is the following:

Try parsing dates using all valid ISO 8601 formats, by calling parse_iso_8601.

If this fails, then try parsing them using the git date parser.

If this fails, then try parsing them using as.POSIXct. (It is unlikely that this step will parse any dates that the first two steps couldn't, but it is still a logical fallback, to make sure that we can parse at least as many dates as as.POSIXct.

parse_date returns quickly in case of empty input elements.

Usage

parse_date(dates, approx = TRUE)

Arguments

dates
A character vector. An error is reported if the function cannot coerce this parameter to a character vector.

approx
Logical flag, whether the git parse should try hard(er). If this is set to TRUE, then the current time is used to fill in the missing parts of the date and time.

Value

A POSIXct vector. NA is returned for the dates that parse_date could not parse.

Examples

# Some easy examples
parse_date("2014-12-12")
parse_date("04/15/99")
parse_date("15/04/99")

# Ambiguous format, parsed assuming MM/DD/YY
parse_date("12/11/99")
parse_date("11/12/99")

# Fill in the current date and time
parse_date("03/20")
parse_date("12")

# But not for this, because this is ISO 8601
parse_date("2014")

# Handle vectors and empty input
parse_date(c("2014","2015","","2016"))

[Package parsedate version 1.1.3 Index]