Parse date from any format

Recognize and parse dates from a wide range of formats. The current algorithm is the following:

Try parsing dates using all valid ISO 8601 formats, by calling parse_iso_8601.

If this fails, then try parsing them using the git date parser.

If this fails, then try parsing them using as.POSIXct. (It is unlikely that this step will parse any dates that the first two steps couldn't, but it is still a logical fallback, to make sure that we can parse at least as many dates as as.POSIXct.

parse_date returns quickly in case of empty input elements.

parse_date(dates, approx = TRUE, default_tz = "UTC")

Arguments

dates

A character vector. An error is reported if the function cannot coerce this parameter to a character vector.

approx

Logical flag, whether the git parse should try hard(er). If this is set to TRUE, then the current time is used to fill in the missing parts of the date and time.

default_tz

Time zone to assume for dates that don't specify a time zone explicitly. Defaults to UTC, and an empty string means the local time zone.

Value

A POSIXct vector. NA is returned for the dates that parse_date could not parse.

Details

All dates are returned in the UTC time zone. If you prefer a different time zone, simply use `.POSIXct()` on the result, see examples below.

Examples

# Some easy examples
parse_date("2014-12-12")
#> [1] "2014-12-12 UTC"
parse_date("04/15/99")
#> [1] "1999-04-15 UTC"
parse_date("15/04/99")
#> [1] "1999-04-15 UTC"

# Ambiguous format, parsed assuming MM/DD/YY
parse_date("12/11/99")
#> [1] "1999-12-11 UTC"
parse_date("11/12/99")
#> [1] "1999-11-12 UTC"

# Fill in the current date and time
parse_date("03/20")
#> [1] "2024-03-20 UTC"
parse_date("12")
#> [1] "2024-12-12 UTC"

# But not for this, because this is ISO 8601
parse_date("2014")
#> [1] "2014-01-01 UTC"

# Handle vectors and empty input
parse_date(c("2014","2015","","2016"))
#> [1] "2014-01-01 UTC" "2015-01-01 UTC" NA               "2016-01-01 UTC"

# Convert result to local time
tz <- format(Sys.time(), "%Z")
as.POSIXct(parse_date("2014-12-13T11:12:13"), tz)
#> [1] "2014-12-13 11:12:13 UTC"

# Local time zone
parse_date("2014-12-13T11:12:13", default_tz = "CET")
#> [1] "2014-12-13 10:12:13 UTC"
parse_date("2014-12-13T11:12:13", default_tz = "UTC")
#> [1] "2014-12-13 11:12:13 UTC"

# Convert results to different timezone
parse_date("2015-12-13T11:12:13")
#> [1] "2015-12-13 11:12:13 UTC"
.POSIXct(parse_date("2015-12-13T11:12:13"), tz = "CET")
#> [1] "2015-12-13 12:12:13 CET"